More additions - Boar user data retrieval/fixing updates - Added multithreading for event listeners   - This allows multiple interactions to be processed at the same time - Added several data util classes to deal with global data - Added JsonUtil to easily convert a file to json string

+++ src/main/java/dev/boarbot/entities/boaruser/collectibles/CollectedPowerup.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 public class CollectedPowerup {
     private long numTotal = 0;
     private long highestTotal = 0;
-    private int numClaimed = 0;
+    private long numClaimed = 0;
     private int numUsed = 0;
     private Integer numActive;
     private Integer numOpened;
